pyproject.toml
[tool.poetry]
name = "pomodoro-time"
version = "0.1.0"
description = ""
authors = ["Pavel Tarakanov <pavel191094@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.1"
uvicorn = "^0.30.6"
pydantic = "^2.9.1"
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
redis = "^5.0.8"
httpx = "^0.27.2"
asyncpg = "^0.29.0"
greenlet = "^3.1.1"
gunicorn = "^23.0.0"
uvicorn-worker = "^0.2.0"
pytest = {extras = ["async"], version = "^8.3.3"}
twisted = "^24.7.0"
anyio = "^4.6.0"
pytest-asyncio = "^0.24.0"
pytest-tornasync = "^0.6.0.post2"
pytest-trio = "^0.8.0"
pytest-twisted = "^1.14.3"
pytest-factoryboy = "^2.7.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
